Initial build passing lint
