+++ .eslintrc
@@ -6,7 +6,8 @@
 	"env": {
 		"browser": true,
 		"node": true,
-		"es6": true
+		"es6": true,
+		"amd": true
 	},
 	"ecmaFeatures": {
 		"jsx": true,
@@ -32,7 +33,7 @@
 		"func-names": [0],
 		"func-style": [0, "declaration"],
 		"generator-star-spacing": [2, "after"],
-		"global-strict": [2, "always"],
+		"strict": [2, "always"],
 		"guard-for-in": [0],
 		"handle-callback-err": [0],
 		"key-spacing": [2, { "beforeColon": false, "afterColon": true }],
@@ -63,7 +64,6 @@
 		"no-duplicate-case": [2],
 		"no-else-return": [0],
 		"no-empty": [2],
-		"no-empty-class": [2],
 		"no-empty-label": [2],
 		"no-eq-null": [0],
 		"no-eval": [2],
@@ -73,7 +73,7 @@
 		"no-extra-boolean-cast": [2],
 		"no-extra-parens": [0],
 		"no-extra-semi": [1],
-		"no-extra-strict": [2],
+		"no-empty-character-class": [2],
 		"no-fallthrough": [2],
 		"no-floating-decimal": [2],
 		"no-func-assign": [2],
@@ -93,7 +93,7 @@
 		"no-multi-spaces": [2],
 		"no-multi-str": [2],
 		"no-multiple-empty-lines": [2, { "max": 2 }],
-		"no-native-reassign": [1],
+		"no-native-reassign": [0],
 		"no-negated-in-lhs": [2],
 		"no-nested-ternary": [0],
 		"no-new": [2],
@@ -104,7 +104,7 @@
 		"no-obj-calls": [2],
 		"no-octal": [2],
 		"no-octal-escape": [2],
-		"no-param-reassign": [2],
+		"no-param-reassign": [0],
 		"no-path-concat": [0],
 		"no-plusplus": [0],
 		"no-process-env": [0],
@@ -120,7 +120,7 @@
 		"no-sequences": [2],
 		"no-shadow": [2],
 		"no-shadow-restricted-names": [2],
-		"no-space-before-semi": [2],
+		"semi-spacing": [2],
 		"no-spaced-func": [2],
 		"no-sparse-arrays": [2],
 		"no-sync": [0],
@@ -138,7 +138,7 @@
 		"no-void": [0],
 		"no-warning-comments": [0, { "terms": ["todo", "fixme", "xxx"], "location": "start" }],
 		"no-with": [2],
-		"no-wrap-func": [2],
+		"no-extra-parens": [2],
 		"one-var": [0],
 		"operator-assignment": [0, "always"],
 		"operator-linebreak": [2, "after"],
