+++ config.json
@@ -77,6 +77,9 @@
 		"hsts": null,
 		"xssProtection": null
 	},
+	"serializers": [
+		"tenso"
+	],
 	"maxBytes": 1048576,
 	"port": 8000,
 	"routes": {},
