+++ lib/middleware.js
@@ -1,18 +1,25 @@
-const path = require("path"),
-	array = require("retsu"),
-	regex = require(path.join(__dirname, "regex.js")),
-	utility = require(path.join(__dirname, "utility.js"));
+"use strict";
 
-let cookie, session, luscaCsp, luscaCsrf, luscaXframe, luscaP3p, luscaHsts, luscaXssProtection, passportAuth,
-	passportInit, passportSession;
+var path = require("path"),
+    array = require("retsu"),
+    regex = require(path.join(__dirname, "regex.js")),
+    utility = require(path.join(__dirname, "utility.js"));
 
-const rateHeaders = [
-	"x-ratelimit-limit",
-	"x-ratelimit-remaining",
-	"x-ratelimit-reset"
-];
+var cookie = undefined,
+    session = undefined,
+    luscaCsp = undefined,
+    luscaCsrf = undefined,
+    luscaXframe = undefined,
+    luscaP3p = undefined,
+    luscaHsts = undefined,
+    luscaXssProtection = undefined,
+    passportAuth = undefined,
+    passportInit = undefined,
+    passportSession = undefined;
 
-function decorate (req, res, next) {
+var rateHeaders = ["x-ratelimit-limit", "x-ratelimit-remaining", "x-ratelimit-reset"];
+
+function decorate(req, res, next) {
 	var obj = req.server.tenso;
 
 	req.protect = false;
@@ -23,7 +30,9 @@ function decorate (req, res, next) {
 		return obj.error(req, res, status, body);
 	};
 
-	res.redirect = function (uri, perm = true) {
+	res.redirect = function (uri) {
+		var perm = arguments.length <= 1 || arguments[1] === undefined ? true : arguments[1];
+
 		return obj.redirect(req, res, uri, undefined, perm);
 	};
 
@@ -38,15 +47,15 @@ function decorate (req, res, next) {
 	next();
 }
 
-function asyncFlag (req, res, next) {
+function asyncFlag(req, res, next) {
 	req.protectAsync = true;
 	next();
 }
 
-function bypass (req, res, next) {
-	let pass = req.server.tenso.config.auth.unprotect.filter(function (i) {
-			return i.test(req.url);
-		}).length > 0;
+function bypass(req, res, next) {
+	var pass = req.server.tenso.config.auth.unprotect.filter(function (i) {
+		return i.test(req.url);
+	}).length > 0;
 
 	if (pass) {
 		req.unprotect = true;
@@ -55,7 +64,7 @@ function bypass (req, res, next) {
 	next();
 }
 
-function csrfWrapper (req, res, next) {
+function csrfWrapper(req, res, next) {
 	if (req.unprotect) {
 		next();
 	} else {
@@ -63,7 +72,7 @@ function csrfWrapper (req, res, next) {
 	}
 }
 
-function guard (req, res, next) {
+function guard(req, res, next) {
 	if (req.parsed.url === "/login" || req.isAuthenticated()) {
 		next();
 	} else {
@@ -71,8 +80,9 @@ function guard (req, res, next) {
 	}
 }
 
-function parse (req, res, next) {
-	let args, type;
+function parse(req, res, next) {
+	var args = undefined,
+	    type = undefined;
 
 	if (regex.body.test(req.method) && req.body !== undefined) {
 		type = req.headers["content-type"];
@@ -98,10 +108,13 @@ function parse (req, res, next) {
 	next();
 }
 
-function keymaster (req, res, next) {
-	let obj = req.server.tenso,
-		authd = req.session && req.isAuthenticated(),
-		method, result, routes, uri;
+function keymaster(req, res, next) {
+	var obj = req.server.tenso,
+	    authd = req.session && req.isAuthenticated(),
+	    method = undefined,
+	    result = undefined,
+	    routes = undefined,
+	    uri = undefined;
 
 	// No authentication, or it's already happened
 	if (!req.protect || !req.protectAsync || authd) {
@@ -141,9 +154,9 @@ function keymaster (req, res, next) {
 	}
 }
 
-function zuul (req, res, next) {
-	let uri = req.parsed.path,
-		protectd = false;
+function zuul(req, res, next) {
+	var uri = req.parsed.path,
+	    protectd = false;
 
 	array.each(req.server.tenso.protect, function (r) {
 		if (r.test(uri)) {
@@ -162,12 +175,12 @@ function zuul (req, res, next) {
 	}
 }
 
-function rate (req, res, next) {
-	let server = req.server,
-		obj = server.tenso,
-		config = server.config.rate,
-		results = obj.rate(req, config.override),
-		good = results.shift();
+function rate(req, res, next) {
+	var server = req.server,
+	    obj = server.tenso,
+	    config = server.config.rate,
+	    results = obj.rate(req, config.override),
+	    good = results.shift();
 
 	rateHeaders.forEach(function (i, idx) {
 		res.setHeader(i, results[idx]);
@@ -180,7 +193,7 @@ function rate (req, res, next) {
 	}
 }
 
-function valid (req, res, next) {
+function valid(req, res, next) {
 	if (req.allow.indexOf(req.method) > -1) {
 		next();
 	} else {
