+++ lib/regex.js
@@ -1,4 +1,6 @@
-const regex = {
+"use strict";
+
+var regex = {
 	body: /POST|PUT|PATCH/i,
 	body_split: /&|=/,
 	collection: /(.*)(\/.*)$/,
