+++ lib/serializers.js
@@ -1,4 +1,6 @@
-function tenso (arg, err, status) {
+"use strict";
+
+function tenso(arg, err, status) {
 	return {
 		data: arg !== null ? arg : null,
 		error: arg === null ? err.message || err || "Something went wrong" : null,
@@ -7,7 +9,7 @@ function tenso (arg, err, status) {
 	};
 }
 
-let serializers = {
+var serializers = {
 	"application/json": tenso,
 	tenso: tenso
 };
