+++ lib/utility.js
@@ -1,38 +1,44 @@
-const path = require("path"),
-	array = require("retsu"),
-	url = require("url"),
-	session = require("express-session"),
-	cookie = require("cookie-parser"),
-	lusca = require("lusca"),
-	uuid = require("tiny-uuid4"),
-	middleware = require(path.join(__dirname, "middleware.js")),
-	regex = require(path.join(__dirname, "regex.js")),
-	passport = require("passport"),
-	BasicStrategy = require("passport-http").BasicStrategy,
-	BearerStrategy = require("passport-http-bearer").Strategy,
-	FacebookStrategy = require("passport-facebook").Strategy,
-	GoogleStrategy = require("passport-google").Strategy,
-	LinkedInStrategy = require("passport-linkedin").Strategy,
-	LocalStrategy = require("passport-local").Strategy,
-	OAuth2Strategy = require("passport-oauth2").Strategy,
-	SAMLStrategy = require("passport-saml").Strategy,
-	TwitterStrategy = require("passport-twitter").Strategy,
-	RedisStore = require("connect-redis")(session);
+"use strict";
 
-function trim (obj) {
+var path = require("path"),
+    array = require("retsu"),
+    url = require("url"),
+    session = require("express-session"),
+    cookie = require("cookie-parser"),
+    lusca = require("lusca"),
+    uuid = require("tiny-uuid4"),
+    middleware = require(path.join(__dirname, "middleware.js")),
+    regex = require(path.join(__dirname, "regex.js")),
+    passport = require("passport"),
+    BasicStrategy = require("passport-http").BasicStrategy,
+    BearerStrategy = require("passport-http-bearer").Strategy,
+    FacebookStrategy = require("passport-facebook").Strategy,
+    GoogleStrategy = require("passport-google").Strategy,
+    LinkedInStrategy = require("passport-linkedin").Strategy,
+    LocalStrategy = require("passport-local").Strategy,
+    OAuth2Strategy = require("passport-oauth2").Strategy,
+    SAMLStrategy = require("passport-saml").Strategy,
+    TwitterStrategy = require("passport-twitter").Strategy,
+    RedisStore = require("connect-redis")(session);
+
+function trim(obj) {
 	return obj.replace(/^(\s+|\t+|\n+)|(\s+|\t+|\n+)$/g, "");
 }
 
-function explode (obj, arg = ",") {
+function explode(obj) {
+	var arg = arguments.length <= 1 || arguments[1] === undefined ? "," : arguments[1];
+
 	return trim(obj).split(new RegExp("\\s*" + arg + "\\s*"));
 }
 
-function escape (arg) {
+function escape(arg) {
 	return arg.replace(/[-[\]{}()*+?.,\\^$|#]/g, "\\$&");
 }
 
-function capitalize (obj, all = false) {
-	let result;
+function capitalize(obj) {
+	var all = arguments.length <= 1 || arguments[1] === undefined ? false : arguments[1];
+
+	var result = undefined;
 
 	if (all) {
 		result = explode(obj, " ").map(capitalize).join(" ");
@@ -43,12 +49,12 @@ function capitalize (obj, all = false) {
 	return result;
 }
 
-function clone (arg) {
+function clone(arg) {
 	return JSON.parse(JSON.stringify(arg));
 }
 
-function coerce (value) {
-	let tmp;
+function coerce(value) {
+	var tmp = undefined;
 
 	if (value === null || value === undefined) {
 		return undefined;
@@ -71,15 +77,15 @@ function coerce (value) {
 	}
 }
 
-function contains (haystack, needle) {
+function contains(haystack, needle) {
 	return haystack.indexOf(needle) > -1;
 }
 
-function isEmpty (obj) {
+function isEmpty(obj) {
 	return trim(obj) === "";
 }
 
-function iterate (obj, fn) {
+function iterate(obj, fn) {
 	if (obj instanceof Object) {
 		Object.keys(obj).forEach(function (i) {
 			fn.call(obj, obj[i], i);
@@ -89,7 +95,7 @@ function iterate (obj, fn) {
 	}
 }
 
-function merge (a, b) {
+function merge(a, b) {
 	if (a instanceof Object && b instanceof Object) {
 		Object.keys(b).forEach(function (i) {
 			if (a[i] instanceof Object && b[i] instanceof Object) {
@@ -109,18 +115,19 @@ function merge (a, b) {
 	return a;
 }
 
-function auth (obj, config) {
-	let ssl = config.ssl.cert && config.ssl.key,
-		proto = "http" + (ssl ? "s" : ""),
-		realm = proto + "://" + (config.hostname === "localhost" ? "127.0.0.1" : config.hostname) + (config.port !== 80 && config.port !== 443 ? ":" + config.port : ""),
-		async = (config.auth.facebook.enabled || config.auth.google.enabled || config.auth.linkedin.enabled || config.auth.twitter.enabled) !== false,
-		stateless = (config.auth.basic.enabled || config.auth.bearer.enabled) !== false,
-		stateful = (async || config.auth.local.enabled || config.security.csrf) !== false,
-		authMap = {},
-		authUris = [],
-		keys, sesh;
+function auth(obj, config) {
+	var ssl = config.ssl.cert && config.ssl.key,
+	    proto = "http" + (ssl ? "s" : ""),
+	    realm = proto + "://" + (config.hostname === "localhost" ? "127.0.0.1" : config.hostname) + (config.port !== 80 && config.port !== 443 ? ":" + config.port : ""),
+	    async = (config.auth.facebook.enabled || config.auth.google.enabled || config.auth.linkedin.enabled || config.auth.twitter.enabled) !== false,
+	    stateless = (config.auth.basic.enabled || config.auth.bearer.enabled) !== false,
+	    stateful = (async || config.auth.local.enabled || config.security.csrf) !== false,
+	    authMap = {},
+	    authUris = [],
+	    keys = undefined,
+	    sesh = undefined;
 
-	function init (sess) {
+	function init(sess) {
 		middleware.passportInit = passport.initialize();
 		obj.server.use(middleware.passportInit).blacklist(middleware.passportInit);
 
@@ -130,7 +137,7 @@ function auth (obj, config) {
 		}
 	}
 
-	function redirect (req, res) {
+	function redirect(req, res) {
 		res.redirect(config.auth.redirect);
 	}
 
@@ -185,7 +192,7 @@ function auth (obj, config) {
 		obj.server.use(middleware.bypass).blacklist(middleware.bypass);
 
 		if (config.security.csrf) {
-			middleware.luscaCsrf = lusca.csrf({key: config.security.key, secret: config.security.secret});
+			middleware.luscaCsrf = lusca.csrf({ key: config.security.key, secret: config.security.secret });
 			obj.server.use(middleware.csrfWrapper).blacklist(middleware.csrfWrapper);
 		}
 	}
@@ -239,7 +246,7 @@ function auth (obj, config) {
 			}
 
 			(function () {
-				let r = "(?!/auth/(";
+				var r = "(?!/auth/(";
 
 				array.each(authUris, function (i) {
 					r += i.replace("_uri", "") + "|";
@@ -249,14 +256,14 @@ function auth (obj, config) {
 
 				obj.server.use(r, middleware.rate).blacklist(middleware.rate);
 				obj.server.use(r, middleware.guard).blacklist(middleware.guard);
-			}());
+			})();
 
 			config.routes.get["/login"] = config.auth.local.enabled ? keys ? {
 				login_uri: "/auth",
 				instruction: "POST 'username' & 'password' to authenticate"
-			} : {instruction: "POST 'username' & 'password' to authenticate"} : {login_uri: "/auth"};
+			} : { instruction: "POST 'username' & 'password' to authenticate" } : { login_uri: "/auth" };
 		} else if (config.auth.local.enabled) {
-			config.routes.get["/login"] = {instruction: "POST 'username' & 'password' to authenticate"};
+			config.routes.get["/login"] = { instruction: "POST 'username' & 'password' to authenticate" };
 		}
 
 		config.routes.get["/logout"] = function (req, res) {
@@ -270,9 +277,9 @@ function auth (obj, config) {
 
 	if (config.auth.basic.enabled) {
 		(function () {
-			let x = {};
+			var x = {};
 
-			function validate (arg, cb) {
+			function validate(arg, cb) {
 				if (x[arg]) {
 					cb(null, x[arg]);
 				} else {
@@ -281,10 +288,10 @@ function auth (obj, config) {
 			}
 
 			array.each(config.auth.basic.list || [], function (i) {
-				let args = i.split(":");
+				var args = i.split(":");
 
 				if (args.length > 0) {
-					x[args[0]] = {password: args[1]};
+					x[args[0]] = { password: args[1] };
 				}
 			});
 
@@ -303,7 +310,7 @@ function auth (obj, config) {
 				});
 			}));
 
-			middleware.passportAuth = passport.authenticate("basic", {session: stateful});
+			middleware.passportAuth = passport.authenticate("basic", { session: stateful });
 
 			if (async || config.auth.local.enabled) {
 				obj.server.get("/auth/basic", middleware.passportAuth).blacklist(middleware.passportAuth);
@@ -311,14 +318,14 @@ function auth (obj, config) {
 			} else {
 				obj.server.use(middleware.passportAuth).blacklist(middleware.passportAuth);
 			}
-		}());
+		})();
 	}
 
 	if (config.auth.bearer.enabled) {
 		(function () {
-			let x = config.auth.bearer.tokens || [];
+			var x = config.auth.bearer.tokens || [];
 
-			function validate (arg, cb) {
+			function validate(arg, cb) {
 				if (array.contains(x, arg)) {
 					cb(null, arg);
 				} else {
@@ -337,11 +344,11 @@ function auth (obj, config) {
 						return done(null, false);
 					}
 
-					return done(null, user, {scope: "read"});
+					return done(null, user, { scope: "read" });
 				});
 			}));
 
-			middleware.passportAuth = passport.authenticate("bearer", {session: stateful});
+			middleware.passportAuth = passport.authenticate("bearer", { session: stateful });
 
 			if (async || config.auth.local.enabled) {
 				obj.server.get("/auth/bearer", middleware.passportAuth).blacklist(middleware.passportAuth);
@@ -349,7 +356,7 @@ function auth (obj, config) {
 			} else {
 				obj.server.use(middleware.passportAuth).blacklist(middleware.passportAuth);
 			}
-		}());
+		})();
 	}
 
 	if (config.auth.facebook.enabled) {
@@ -371,7 +378,7 @@ function auth (obj, config) {
 		obj.server.get("/auth/facebook", middleware.asyncFlag);
 		obj.server.get("/auth/facebook", passport.authenticate("facebook"));
 		obj.server.get("/auth/facebook/callback", middleware.asyncFlag);
-		obj.server.get("/auth/facebook/callback", passport.authenticate("facebook", {failureRedirect: "/login"}));
+		obj.server.get("/auth/facebook/callback", passport.authenticate("facebook", { failureRedirect: "/login" }));
 		obj.server.get("/auth/facebook/callback", redirect);
 	}
 
@@ -393,7 +400,7 @@ function auth (obj, config) {
 		obj.server.get("/auth/google", middleware.asyncFlag);
 		obj.server.get("/auth/google", passport.authenticate("google"));
 		obj.server.get("/auth/google/callback", middleware.asyncFlag);
-		obj.server.get("/auth/google/callback", passport.authenticate("google", {failureRedirect: "/login"}));
+		obj.server.get("/auth/google/callback", passport.authenticate("google", { failureRedirect: "/login" }));
 		obj.server.get("/auth/google/callback", redirect);
 	}
 
@@ -414,9 +421,9 @@ function auth (obj, config) {
 		}));
 
 		obj.server.get("/auth/linkedin", middleware.asyncFlag);
-		obj.server.get("/auth/linkedin", passport.authenticate("linkedin", {"scope": config.auth.linkedin.scope || ["r_basicprofile", "r_emailaddress"]}));
+		obj.server.get("/auth/linkedin", passport.authenticate("linkedin", { "scope": config.auth.linkedin.scope || ["r_basicprofile", "r_emailaddress"] }));
 		obj.server.get("/auth/linkedin/callback", middleware.asyncFlag);
-		obj.server.get("/auth/linkedin/callback", passport.authenticate("linkedin", {failureRedirect: "/login"}));
+		obj.server.get("/auth/linkedin/callback", passport.authenticate("linkedin", { failureRedirect: "/login" }));
 		obj.server.get("/auth/linkedin/callback", redirect);
 	}
 
@@ -434,7 +441,7 @@ function auth (obj, config) {
 
 		config.routes.post = config.routes.post || {};
 		config.routes.post["/login"] = function (req, res) {
-			function final () {
+			function final() {
 				passport.authenticate("local")(req, res, function (e) {
 					if (e) {
 						res.error(401, "Unauthorized");
@@ -446,7 +453,7 @@ function auth (obj, config) {
 				});
 			}
 
-			function mid () {
+			function mid() {
 				middleware.passportSession(req, res, final);
 			}
 
@@ -475,13 +482,13 @@ function auth (obj, config) {
 		obj.server.get("/auth/oauth2", middleware.asyncFlag);
 		obj.server.get("/auth/oauth2", passport.authenticate("oauth2"));
 		obj.server.get("/auth/oauth2/callback", middleware.asyncFlag);
-		obj.server.get("/auth/oauth2/callback", passport.authenticate("oauth2", {failureRedirect: "/login"}));
+		obj.server.get("/auth/oauth2/callback", passport.authenticate("oauth2", { failureRedirect: "/login" }));
 		obj.server.get("/auth/oauth2/callback", redirect);
 	}
 
 	if (config.auth.saml.enabled) {
 		(function () {
-			let arg = config.auth.saml;
+			var arg = config.auth.saml;
 
 			arg.callbackURL = realm + "/auth/saml/callback";
 			delete arg.enabled;
@@ -497,12 +504,12 @@ function auth (obj, config) {
 					done(null, user);
 				});
 			}));
-		}());
+		})();
 
 		obj.server.get("/auth/saml", middleware.asyncFlag);
 		obj.server.get("/auth/saml", passport.authenticate("saml"));
 		obj.server.get("/auth/saml/callback", middleware.asyncFlag);
-		obj.server.get("/auth/saml/callback", passport.authenticate("saml", {failureRedirect: "/login"}));
+		obj.server.get("/auth/saml/callback", passport.authenticate("saml", { failureRedirect: "/login" }));
 		obj.server.get("/auth/saml/callback", redirect);
 	}
 
@@ -534,8 +541,8 @@ function auth (obj, config) {
 	return config;
 }
 
-function bootstrap (obj, config) {
-	let notify = false;
+function bootstrap(obj, config) {
+	var notify = false;
 
 	// Bootstrapping configuration
 	auth(obj, config);
@@ -553,7 +560,11 @@ function bootstrap (obj, config) {
 	iterate(config.routes, function (routes, method) {
 		iterate(routes, function (arg, route) {
 			if (typeof arg === "function") {
-				obj.server[method](route, function (...args) {
+				obj.server[method](route, function () {
+					for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+						args[_key] = arguments[_key];
+					}
+
 					arg.apply(obj, args);
 				});
 			} else {
@@ -572,8 +583,8 @@ function bootstrap (obj, config) {
 
 	// Starting API server
 	obj.server.start(config, function (req, res, status, msg) {
-		let stat = status instanceof Error ? parseInt(status.message, 10) : status,
-			err = msg instanceof Error ? msg : new Error(msg || obj.messages[stat]);
+		var stat = status instanceof Error ? parseInt(status.message, 10) : status,
+		    err = msg instanceof Error ? msg : new Error(msg || obj.messages[stat]);
 
 		obj.error(req, res, stat, err);
 	});
@@ -585,10 +596,12 @@ function bootstrap (obj, config) {
 	return obj;
 }
 
-function queryString (qstring = "") {
-	let obj = {};
-	let aresult = qstring.split("?");
-	let result;
+function queryString() {
+	var qstring = arguments.length <= 0 || arguments[0] === undefined ? "" : arguments[0];
+
+	var obj = {};
+	var aresult = qstring.split("?");
+	var result = undefined;
 
 	if (aresult.length > 1) {
 		aresult.shift();
@@ -596,8 +609,8 @@ function queryString (qstring = "") {
 
 	result = aresult.join("?");
 	result.split("&").forEach(function (prop) {
-		let aitem = prop.replace(/\+/g, " ").split("=");
-		let item;
+		var aitem = prop.replace(/\+/g, " ").split("=");
+		var item = undefined;
 
 		if (aitem.length > 2) {
 			item = [aitem.shift(), aitem.join("=")];
@@ -628,9 +641,11 @@ function queryString (qstring = "") {
 	return obj;
 }
 
-function parse (uri) {
-	let luri = uri;
-	let idxAscii, idxQ, parsed;
+function parse(uri) {
+	var luri = uri;
+	var idxAscii = undefined,
+	    idxQ = undefined,
+	    parsed = undefined;
 
 	if (luri === undefined || luri === null) {
 		luri = "";
@@ -660,22 +675,28 @@ function parse (uri) {
 	return parsed;
 }
 
-function hypermedia (server, req, rep, headers) {
-	let seen = {},
-		collection = req.parsed.pathname,
-		query, page, page_size, nth, root, parent;
+function hypermedia(server, req, rep, headers) {
+	var seen = {},
+	    collection = req.parsed.pathname,
+	    query = undefined,
+	    page = undefined,
+	    page_size = undefined,
+	    nth = undefined,
+	    root = undefined,
+	    parent = undefined;
 
 	// Parsing the object for hypermedia properties
-	function marshal (obj, rel, item_collection) {
-		let keys = array.keys(obj),
-			lrel = rel || "related",
-			result;
+	function marshal(obj, rel, item_collection) {
+		var keys = array.keys(obj),
+		    lrel = rel || "related",
+		    result = undefined;
 
 		if (keys.length === 0) {
 			result = null;
 		} else {
 			array.each(keys, function (i) {
-				let lcollection, uri;
+				var lcollection = undefined,
+				    uri = undefined;
 
 				// If ID like keys are found, and are not URIs, they are assumed to be root collections
 				if (regex.id.test(i) || regex.hypermedia.test(i)) {
@@ -693,7 +714,7 @@ function hypermedia (server, req, rep, headers) {
 						seen[uri] = 1;
 
 						if (server.allowed("get", uri, req.vhost)) {
-							rep.links.push({uri: uri, rel: lrel});
+							rep.links.push({ uri: uri, rel: lrel });
 						}
 					}
 				}
@@ -732,40 +753,40 @@ function hypermedia (server, req, rep, headers) {
 					query.page_size = page_size;
 
 					root += "?" + array.keys(query).map(function (i) {
-							return i + "=" + encodeURIComponent(query[i]);
-						}).join("&");
+						return i + "=" + encodeURIComponent(query[i]);
+					}).join("&");
 
 					if (page > 1) {
-						rep.links.push({uri: root.replace("page=0", "page=1"), rel: "first"});
+						rep.links.push({ uri: root.replace("page=0", "page=1"), rel: "first" });
 					}
 
 					if (page - 1 > 1 && page <= nth) {
-						rep.links.push({uri: root.replace("page=0", "page=" + (page - 1)), rel: "prev"});
+						rep.links.push({ uri: root.replace("page=0", "page=" + (page - 1)), rel: "prev" });
 					}
 
 					if (page + 1 < nth) {
-						rep.links.push({uri: root.replace("page=0", "page=" + (page + 1)), rel: "next"});
+						rep.links.push({ uri: root.replace("page=0", "page=" + (page + 1)), rel: "next" });
 					}
 
 					if (nth > 0 && page !== nth) {
-						rep.links.push({uri: root.replace("page=0", "page=" + nth), rel: "last"});
+						rep.links.push({ uri: root.replace("page=0", "page=" + nth), rel: "last" });
 					}
 				} else {
 					root += "?" + array.keys(query).map(function (i) {
-							return i + "=" + encodeURIComponent(query[i]);
-						}).join("&");
+						return i + "=" + encodeURIComponent(query[i]);
+					}).join("&");
 				}
 			}
 
 			array.each(rep.data, function (i) {
 				var li = i.toString(),
-					uri;
+				    uri;
 
 				if (li !== collection) {
 					uri = li.indexOf("//") > -1 || li.indexOf("/") === 0 ? li : (collection + "/" + li).replace(/^\/\//, "/");
 
 					if (server.allowed("get", uri, req.vhost)) {
-						rep.links.push({uri: uri, rel: "item"});
+						rep.links.push({ uri: uri, rel: "item" });
 					}
 				}
