+++ package.json
@@ -23,7 +23,6 @@
     "connect-redis": "^3.0.1",
     "cookie-parser": "^1.4.0",
     "express-session": "^1.12.1",
-    "keigai": "~1.3.19",
     "lusca": "^1.3.0",
     "passport": "^0.3.2",
     "passport-facebook": "^2.0.0",
@@ -47,16 +46,13 @@
     "babel-preset-es2015": "^6.1.2",
     "grunt": "^0.4.5",
     "grunt-babel": "^6.0.0",
-    "grunt-cli": "^0.1.13",
-    "grunt-contrib-concat": "^0.1.3",
     "grunt-contrib-sass": "^0.9.2",
     "grunt-contrib-watch": "^0.2.0",
     "grunt-eslint": "^17.3.1",
-    "grunt-jsdoc": "~0.5.6",
     "grunt-mocha-test": "^0.12.7",
     "grunt-nsp": "^2.1.2",
+    "grunt-sed": "^0.1.1",
     "hippie": "^0.4.0",
-    "ink-docstrap": "~0.4.12",
     "mocha": "^2.3.2"
   },
   "keywords": [
